Update list page

+++ src/js/components/ProjectListContainer/index.jsx
@@ -2,6 +2,7 @@
 import React from 'react'
 import classnames from 'classnames'
 import style from './style.scss'
+import projectListStyle from '../shared/List/ProjectList/style.scss'
 import HttpUtil from '../../utils/HttpUtil'
 import ProjectList from '../shared/List/ProjectList'
 import TextFieldWithButton from '../shared/TextFieldWithButton'
@@ -94,9 +95,7 @@ export default class ProjectListContainer extends React.Component {
 
   renderSearchBar () {
     return <div className={style.search_bar}>
-      <TextFieldWithButton placeholder={'プロジェクトを検索'}
-                           onChange={(e) => this.onChangeKeyword(
-                             e)}>検索</TextFieldWithButton>
+      <TextField placeholder={'プロジェクトを検索'} onChange={(e) => this.onChangeKeyword(e)}/>
     </div>
   }
 
@@ -128,8 +127,21 @@ export default class ProjectListContainer extends React.Component {
 
   onClickDelete (project_uuid) {
     const self = this
-    HttpUtil.delete('projects/' + project_uuid).then((response) => {
-      self.getProjectList()
+    ModalUtil.registerModal({
+      id: Constants.modal.CONFIRM, onClickDone: () => {
+        HttpUtil.delete('projects/' + project_uuid).then((response) => {
+          self.getProjectList()
+          ModalUtil.closeModal(Constants.modal.CONFIRM)
+        })
+      },
+    })
+    ModalUtil.emitModal({
+      id: Constants.modal.CONFIRM,
+      visible: true,
+      done: '削除する',
+      content: <div>
+        選択されたプロジェクトを削除しますか？
+      </div>,
     })
   }
 
@@ -143,7 +155,14 @@ export default class ProjectListContainer extends React.Component {
   }
 
   renderNewProject () {
-    return <div className={"mt-20px"}><a href="#" onClick={(e) => this.onClickNew(e)}>新しくプロジェクトを作成する</a></div>
+    return <a className={classnames(projectListStyle.project,projectListStyle.new)} href="#" onClick={(e) => this.onClickNew(e)}>
+      <div className={projectListStyle.project_list}>
+        <div className={projectListStyle.name}>
+          <i className={classnames('material-icons', [projectListStyle.icon])}>add_circle_outline</i>
+          新しくプロジェクトを作成する
+        </div>
+      </div>
+    </a>
   }
 
   renderAll () {
